snapshot

+++ v2/runner.py
@@ -4,19 +4,24 @@ import enum
 import json
 import multiprocessing.dummy
 import os
+import pickle
 import shutil
 import threading
 import time
+import traceback
 from typing import Deque, Dict, Generic, Iterable, List, Optional, Set, TypeVar
 
+from torch.utils.benchmark import CallgrindStats, Language, Measurement
+
 from v2.build import build_clean
+from v2.containers import BenchmarkResult, BenchmarkResults
 from v2 import init_pytorch
-from v2.runner_interface import backport, clean_backport, get_benchmarks, BenchmarkRunner, CorePool, WorkOrder
+from v2.runner_interface import backport, clean_backport, get_benchmarks, BenchmarkRunner, InstrumentedCorePool, RuntimeMode, WorkOrder
 from v2.logging_subprocess import call
-from v2.runner_policy import BASELINE_BUILD_CORE_COUNT, NUM_CORES, OPPORTUNISTIC_BUILD_CORE_COUNTS, POOL_SLACK
+from v2.runner_policy import benchmark_may_fail, BASELINE_BUILD_CORE_COUNT, MAX_BUILD_CORES, NUM_CORES, OPPORTUNISTIC_BUILD_CORE_COUNTS, POOL_SLACK
 from v2.workspace import (
-    BUILD_COMPLETED_ROOT, DATE_FMT, MUTATION_LOCK, RUN_LOG_ROOT, RUNNER_STATE_ROOT,
-    STOP_FILE, SWEEP_CADENCE, SWEEP_START)
+    BUILD_COMPLETED_ROOT, DATE_FMT, MUTATION_LOCK, PDB_FILE, RUN_LOG_ROOT,
+    RUN_COMPLETED_ROOT, RUNNER_STATE_ROOT, STOP_FILE, SWEEP_CADENCE, SWEEP_START)
 
 
 _STATE_PATH = os.path.join(RUNNER_STATE_ROOT, "state.json")
@@ -192,7 +197,9 @@ class Runner:
         MUTATION_LOCK.get()
         # init_pytorch.fetch_fbcode_warm()
         self._state = RunnerState.from_file(frozen=False)
-        self._core_pool = CorePool(POOL_SLACK)
+        self._core_pool = InstrumentedCorePool(POOL_SLACK)
+        self._remaining_build_cores = MAX_BUILD_CORES
+        self._lock = threading.RLock()
         self._history = init_pytorch.get_history().since(SWEEP_START)
         self._history_dict = {commit.sha: commit for commit in self._history}
 
@@ -206,6 +213,7 @@ class Runner:
 
         # Debug:
         self._initial_sweep_indices = self._initial_sweep_indices[:100]
+        # self._initial_sweep_indices = self._initial_sweep_indices[:10]
 
         for i in self._initial_sweep_indices:
             self._state.maybe_enqueue_build(self._history[i].sha)
@@ -214,29 +222,46 @@ class Runner:
         self._benchmarks = None
 
     def run(self):
-        # assert self._benchmarks is None
-        # self._benchmarks = get_benchmarks()
-        # self.measure()
+        assert self._benchmarks is None
+        self._benchmarks = get_benchmarks()
 
-        # self.worker_fn(0)
-        num_workers = 8  # TODO: tune.
+        # Does not include control loop.
+        num_workers = 6  # TODO: tune.
         with multiprocessing.dummy.Pool(num_workers) as pool:
-            pool.map(self.worker_fn, range(num_workers), 1)
+            pool.map(self.worker_fn, range(num_workers + 1), 1)
 
     def worker_fn(self, worker_id: int):
-        # TODO: bisection and test loop will have higher priority.
+        # Worker 0 is the control loop.
+        if not worker_id:
+            while not self._stop:
+                if os.path.exists(PDB_FILE):
+                    os.remove(PDB_FILE)
+                    import pdb
+                    pdb.set_trace()
+
+                    # print(sum([n.num_available for n in self._core_pool._nodes]))
+
+                if os.path.exists(STOP_FILE):
+                    os.remove(STOP_FILE)
+                    self._stop = True
+                    return
+
+                if not self.scheduled_work:
+                    self._stop = True  # Debug
+
+                # TODO: Bisection.
+
+                time.sleep(1)
+            return
 
         while not self._stop:
-            if os.path.exists(STOP_FILE):
-                os.remove(STOP_FILE)
-                self._stop = True
-                break
+            if worker_id <= 3:
+                # Reserve three workers for building.
+                self.build()
 
-            if not self.scheduled_work:
-                self._stop = True  # Debug
-
-            # self.measure()
-            self.build()
+            else:
+                self.measure()
+                self.build()
             time.sleep(10)
 
     @property
@@ -249,44 +274,53 @@ class Runner:
         ])
 
     def build(self):
-        sha = self._state.get_job(TaskType.BUILD)
-        if sha is None:
-            return
+        with self._lock:
+            if not self._state.queue_len(TaskType.BUILD):
+                return
 
-        commit = self._history_dict[sha]
+            priority = (BASELINE_BUILD_CORE_COUNT,)
+            if self.scheduled_work + BASELINE_BUILD_CORE_COUNT < NUM_CORES:
+                priority = OPPORTUNISTIC_BUILD_CORE_COUNTS + priority
 
-        priority = (BASELINE_BUILD_CORE_COUNT,)
-        if self.scheduled_work + BASELINE_BUILD_CORE_COUNT < NUM_CORES:
-            priority = OPPORTUNISTIC_BUILD_CORE_COUNTS + priority
+            for core_request in priority:
+                if core_request > self._remaining_build_cores:
+                    continue
 
-        for core_request in priority:
-            allocation = self._core_pool.reserve(core_request)
-            if allocation:
-                break
+                allocation = self._core_pool.reserve(core_request)
+                if allocation:
+                    self._remaining_build_cores -= core_request
+                    break
 
-        if allocation is None:
-            return
+            if allocation is None:
+                return
 
-        print(f"Building: {sha} ({commit.date_str})  {allocation}")
-        start_time = time.time()
-        conda_env = build_clean(
-            sha,
-            commit.build_cfg,
-            show_progress=False,
-            taskset_cores=allocation,
-            nice="15",
-            max_jobs=core_request,
-        )
+            sha = self._state.get_job(TaskType.BUILD)
+            assert sha is not None
 
-        self._core_pool.release(allocation)
+        try:
+            commit = self._history_dict[sha]
+            print(f"Building: {sha} ({commit.date_str})  {allocation}")
+            start_time = time.time()
+            conda_env = build_clean(
+                sha,
+                commit.build_cfg,
+                show_progress=False,
+                taskset_cores=allocation,
+                nice="15",
+                max_jobs=core_request,
+            )
 
-        if conda_env is None:
-            return  # unbuildable
+            if conda_env is None:
+                return  # unbuildable
 
-        new_env_path = os.path.join(BUILD_COMPLETED_ROOT, f"{commit.date_str}_{sha}")
-        shutil.move(conda_env, new_env_path)
-        self._state.report_finished(TaskType.BUILD, sha, new_env_path)
-        print(f"Build time: {sha} ({commit.date_str}) {time.time() - start_time:.0f} sec")
+            new_env_path = os.path.join(BUILD_COMPLETED_ROOT, f"{commit.date_str}_{sha}")
+            shutil.move(conda_env, new_env_path)
+            self._state.report_finished(TaskType.BUILD, sha, new_env_path)
+            print(f"Build time: {sha} ({commit.date_str}) {time.time() - start_time:.0f} sec")
+
+        finally:
+            self._core_pool.release(allocation)
+            self._remaining_build_cores += core_request
 
     def measure(self):
         sha = self._state.get_job(TaskType.TEST)
@@ -311,44 +345,101 @@ class Runner:
         clean_backport(torch_path)
 
         try:
+            # Valgrind will sometimes fail with:
+            #   `valgrind: failed to start tool 'callgrind' for platform 'amd64-linux': No such file or directory`
+            # but a re-install fixes it.
+            check_valgrind_retcode = call(
+                "valgrind --version",
+                shell=True,
+                conda_env=conda_env,
+                timeout=60,
+                task_name="Check Valgrind",
+                log_dir=RUN_LOG_ROOT,
+            )
+
+            if check_valgrind_retcode:
+                call(
+                    """
+                    conda remove -y valgrind
+                    conda install -y valgrind -c conda-forge
+                    valgrind --version
+                    """,
+                    shell=True,
+                    check=True,
+                    conda_env=conda_env,
+                    timeout=60,
+                    task_name="Reinstall Valgrind",
+                    log_dir=RUN_LOG_ROOT,
+                )
+
             backport(torch_path)
-            # retcode = call(
-            #     """
-            #     conda remove -y valgrind
-            #     conda install -y valgrind -c conda-forge
-            #     python -c "from torch.utils_backport.benchmark import Timer; print(Timer('torch::empty({1});', language='c++').collect_callgrind(repeats=10))"
-            #     """,
-            #     shell=True,
-            #     # check=True,
-            #     conda_env=conda_env,
-            #     # per_line_fn=lambda l: print(l),
-            #     log_dir=RUN_LOG_ROOT,
-            # )
-            # if retcode:
-            #     print(f"Failed:  {sha}")
-            # else:
-            #     print(f"Success: {sha}")
 
             work_orders: List[WorkOrder] = []
-            for label, auto_labels, timer_args in self._benchmarks[:10]:
-                # TODO: Figure out when jit models will work. (And fix enum compat.)
-                if auto_labels.runtime.value == "TorchScript":
-                    continue
+            for label, auto_labels, timer_args in self._benchmarks:
                 work_orders.append(WorkOrder(
                     label=label,
                     auto_labels=auto_labels,
                     timer_args=timer_args,
                     source_cmd=f"source activate {conda_env}",
-                    timeout=300.0,
+                    timeout=360.0,
                     retries=3,
+                    allow_failure=benchmark_may_fail(
+                        is_cpp=timer_args.language == Language.CPP,
+                        is_jit=auto_labels.runtime == RuntimeMode.JIT,
+                        label=label,
+                    ),
                 ))
 
+            benchmark_runner = BenchmarkRunner(
+                work_items=tuple(work_orders),
+                core_pool=self._core_pool,
+                display_progress=False,
+            )
 
-            results = BenchmarkRunner(work_items=tuple(work_orders), core_pool=self._core_pool, display_progress=True).run()
-            import pdb
-            pdb.set_trace()
+            try:
+                self._core_pool.register_benchmark_runner(benchmark_runner)
+                results = benchmark_runner.run()
+                simple_results = []
+                raw_results = []
+                lang_to_str = {Language.PYTHON: "Python", Language.CPP: "C++"}
+                num_failures = 0
+                for w in work_orders:
+                    if w not in results:
+                        num_failures += 1
+                        continue
 
-            print(conda_env)
+                    r = results[w]
+                    simple_results.append(BenchmarkResult(
+                        lang_to_str[w.auto_labels.language],
+                        w.auto_labels.autograd.value,
+                        w.auto_labels.runtime.value,
+                        r.wall_time.median,
+                        r.instructions.counts(denoise=True),
+                    ))
+                    raw_results.append((w.auto_labels, r))
+
+                simple_path = os.path.join(RUN_COMPLETED_ROOT, f"{commit.date_str}__{sha}.pkl")
+                raw_path = os.path.join(RUN_COMPLETED_ROOT, f"{commit.date_str}__{sha}_raw.pkl")
+
+                with open(simple_path, "wb") as f:
+                    pickle.dump(BenchmarkResults(
+                        sha=sha,
+                        conda_env=conda_env,
+                        values=tuple(simple_results)
+                    ), f)
+
+                with open(raw_path, "wb") as f:
+                    pickle.dump([sha, conda_env, raw_results], f)
+
+                self._state.report_finished(TaskType.TEST, sha=sha, result=simple_path)
+                print(f"Test finished: {sha} {commit.date_str}  {num_failures} / {len(work_orders)} failures")
+
+
+            except BaseException as e:
+                print(f"Failed: {sha} {e}")
+
+            finally:
+                self._core_pool.unregister_benchmark_runner()
+
         finally:
-            pass
-            # clean_backport(torch_path)
+            clean_backport(torch_path)
